edit home

+++ src/app/recipes/recipe-detail/recipe-detail.component.ts
@@ -3,6 +3,7 @@ import { Recipe } from '../recipe.model';
 import { RecipeService } from '../recipe.service';
 import { ActivatedRoute, Params, Router } from '@angular/router';
 import { Subscription } from 'rxjs';
+import { AuthService } from 'src/app/auth/auth.service';
 
 @Component({
   selector: 'app-recipe-detail',
@@ -12,33 +13,36 @@ import { Subscription } from 'rxjs';
 export class RecipeDetailComponent implements OnInit, OnDestroy {
   recipe?: Recipe;
   id?: number
-  subscription:Subscription
+  subscription: Subscription
   constructor(private recipeService: RecipeService,
-              private route: ActivatedRoute,
-              private router: Router) { }
-  
+    private route: ActivatedRoute,
+    private router: Router,
+    private authService: AuthService) { }
 
-  ngOnInit(){
+
+  ngOnInit() {
     this.subscription = this.route.params.subscribe(
       (params: Params) => {
         this.id = +params['id'];
         this.recipe = this.recipeService.getRecipeById(this.id)
       }
     )
-    
-    
+
+
   }
 
   onAddToShoppingList() {
-    if (this.recipe)
+    if (this.recipe && this.authService.isAuthenticated())
       this.recipeService.addIngredientsToShoppingList(this.recipe.ingredients)
   }
 
   onDelete() {
-    this.recipeService.deleteRecipe(this.id);
-    this.router.navigate(['recipes'])
+    if (this.authService.isAuthenticated()) {
+      this.recipeService.deleteRecipe(this.id);
+      this.router.navigate(['recipes'])
+    }
   }
-  
+
   ngOnDestroy(): void {
     this.subscription.unsubscribe()
   }
